requirements/requirements.txt
Django>=5.0.6,<5.1
django-rest-swagger>=2.2.0,<3.0
djangorestframework>=3.15.1,<3.16
drf-yasg>=1.21.7,<2.0
whitenoise>=6.6.0,<7.0
psycopg2>=2.9.9,<3.0
django-cors-headers>=4.3.1,<5.0
django-redis>=5.4.0,<6.0
django-environ>=0.11.2,<0.12
